# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Universal SOP Architect",
    "description": "专家经验数字化编译平台 - 将业务专家的直觉与流程编译为 AI 可执行的协议",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/steps/{step_id}/examples": {
      "get": {
        "tags": [
          "examples"
        ],
        "summary": "List Examples",
        "description": "List all examples for a step.\n\n- **step_id**: UUID of the step\n- **label**: Optional filter by label (PASS or FAIL)",
        "operationId": "list_examples_api_steps__step_id__examples_get",
        "parameters": [
          {
            "name": "step_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Step Id"
            }
          },
          {
            "name": "label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(PASS|FAIL)$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by label (PASS or FAIL)",
              "title": "Label"
            },
            "description": "Filter by label (PASS or FAIL)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExampleResponse"
                  },
                  "title": "Response List Examples Api Steps  Step Id  Examples Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "examples"
        ],
        "summary": "Create Example",
        "description": "Create a new example for a step.\n\n- **step_id**: UUID of the step\n- **content**: Example content (URL for image, text for text)\n- **content_type**: Content type ('image' or 'text')\n- **label**: Label ('PASS' or 'FAIL') - determines which upload zone\n- **description**: Optional description\n\nNote: The label determines whether this is a passing or failing example.\nExamples uploaded to the passing zone should have label='PASS',\nand examples uploaded to the failing zone should have label='FAIL'.",
        "operationId": "create_example_api_steps__step_id__examples_post",
        "parameters": [
          {
            "name": "step_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Step Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExampleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/examples/{example_id}": {
      "get": {
        "tags": [
          "examples"
        ],
        "summary": "Get Example",
        "description": "Get an example by ID.\n\n- **example_id**: UUID of the example",
        "operationId": "get_example_api_examples__example_id__get",
        "parameters": [
          {
            "name": "example_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Example Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "examples"
        ],
        "summary": "Update Example",
        "description": "Update an example.\n\n- **example_id**: UUID of the example\n- All fields are optional for partial updates",
        "operationId": "update_example_api_examples__example_id__put",
        "parameters": [
          {
            "name": "example_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Example Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExampleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "examples"
        ],
        "summary": "Delete Example",
        "description": "Delete an example.\n\n- **example_id**: UUID of the example",
        "operationId": "delete_example_api_examples__example_id__delete",
        "parameters": [
          {
            "name": "example_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Example Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/files/upload": {
      "post": {
        "tags": [
          "files"
        ],
        "summary": "Upload a file",
        "description": "Upload an image file. Supported formats: JPG, JPEG, PNG, GIF, WEBP, BMP. Max size: 10MB.",
        "operationId": "upload_file_api_files_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_files_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/delete": {
      "delete": {
        "tags": [
          "files"
        ],
        "summary": "Delete a file",
        "description": "Delete a previously uploaded file by its URL.",
        "operationId": "delete_file_api_files_delete_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_file_api_files_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_files_upload_post"
      },
      "ExampleCreate": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "image",
                  "text"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "label": {
            "type": "string",
            "enum": [
              "PASS",
              "FAIL"
            ],
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "content",
          "label"
        ],
        "title": "ExampleCreate",
        "description": "Schema for creating an Example."
      },
      "ExampleResponse": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "image",
                  "text"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "label": {
            "type": "string",
            "enum": [
              "PASS",
              "FAIL"
            ],
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "step_id": {
            "type": "string",
            "format": "uuid",
            "title": "Step Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "content",
          "label",
          "id",
          "step_id",
          "created_at"
        ],
        "title": "ExampleResponse",
        "description": "Schema for Example response."
      },
      "ExampleUpdate": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "image",
                  "text"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "label": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "PASS",
                  "FAIL"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "ExampleUpdate",
        "description": "Schema for updating an Example."
      },
      "FileDeleteRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "FileDeleteRequest",
        "description": "Request model for file deletion."
      },
      "FileDeleteResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "FileDeleteResponse",
        "description": "Response model for file deletion."
      },
      "FileUploadResponse": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "url",
          "filename",
          "message"
        ],
        "title": "FileUploadResponse",
        "description": "Response model for file upload."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}